+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - SC2321: Suggest removing $((..)) in array[$((idx))]=val
 - SC2322: Suggest collapsing double parentheses in arithmetic contexts
 - SC2323: Suggest removing wrapping parentheses in a[(x+1)]=val
+- support Static Analysis Results Interchange Format (SARIF) output
 
 ### Fixed
 - SC2086: Now uses DFA to make more accurate predictions about values